Add retry mechanism for HCI interface detection
The HCI controller sometimes does not register immediately, leading to HCI interface detection failures. And then, HCI can't be opened successfully.
Add retry mechanism with a 500ms delay to fix AIDL is stuck when polling
hci interface event. And add crash mechanism to fix AIDL is not aware of
EPIPE when read data.

Bug: 362663031
Test: Unplug the wifi chip after Bluetooth state is on. |
    Bluetooth can restart successfully after re-plug it again.
Change-Id: I09f8e4fa6fda445557a3c433eca71363cbb2a047