hal: Make destructors virtual for bluetooth_hal classes
Making destructors virtual in base classes is a C++ best practice. It
ensures that when a derived class object is deleted through a base class
pointer, the destructors of both the base and derived classes are
called, in the correct order. This prevents resource leaks and undefined
behavior that can occur if the derived class destructor is not invoked.

Bug: 440426822
Test: sh run_test.sh
Flag: EXEMPT, simple refactor.
Change-Id: I28f36d69016986226eb2bfd68bce40a03b74d04d