Add connectionCapability to TrafficDescriptor HAL
This change updates the Radio HAL interface to include the
`connectionCapability` field in the `TrafficDescriptor` AIDL
parcelable, allowing the new field to be passed down to the
modem during the `setupDataCall` process. The values are based
on 3GPP TS 24.526.

Test: build
Bug: 446768290
Bug: 417753705
flag: com.android.internal.telephony.flags.enable_traffic_descriptor_connection_capability
Change-Id: I7787a5396c5aa4786706396759183640d06e9fc3