refactor: Group modules in bluetooth_hal/Android.bp
Add comments to group the modules in the Android.bp file for the
bluetooth_hal library.

The file was becoming large and difficult to navigate. This change
improves readability by organizing the modules into logical sections:
- Defaults
- FileGroups
- Libraries
- Tests

This makes it easier to find and understand the different components
of the bluetooth_hal library.

Bug: 447708157
Test: Build passed.
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ia44ece345b617349eb0015087f8d6c0f6f008d65