Deprecate retryAfterDuration in RttResult
Deprecate the `retryAfterDuration` field in `RttResult` and introduce
`retryAfterDurationMillis`. The original `retryAfterDuration` was a
byte,
which was too small to represent a sufficient range of retry durations.
The new field is an int, allowing for a much larger range of values in
milliseconds.

To maintain backward compatibility, the conversion from the legacy HAL
struct continues to populate the old `retryAfterDuration` field as well
as the new `retryAfterDurationMillis` field.

Unit tests and VTS tests have been updated to cover the new field.

Bug: 407453160
Test: atest WifiRttControllerAidlTest
Flag: com.android.wifi.flags.aware_periodic_ranging_interval
Change-Id: I6205adb2dec66ebd98a9f21c0849c28e5890d9d1