Add CTRL function to GadgetFunction AIDL.
This change introduces a new `CTRL` bitmask in `GadgetFunction` to
represent a vendor-specific control interface. It also adds a VTS test
for USB Gadget CTRL interface. This change introduces a new test case to
verify the presence of a vendor-specific CTRL interface (Class=0xFF,
Subclass=0xFF, Protocol=0x00) on the host. The test is only run on
devices with the flag enabled.

Bug: 448528878
Test: Manually tested that the HAL and the framework pick up the new GadgetFunction
Test: VtsHalUsbGadgetV2_0HostTest
Change-Id: I8a64ce2631e30bec88bac8023b30753ca37071cf