Fix LE Audio asymmetric codec info test
The test for asymmetric LE Audio codec info was checking for strict
equality on the transport flags. This is too restrictive, as other
flags may also be present.

This change updates the test to check only for the required flag's
bitmask, making the test more robust to other flags being set.

Test: atest BluetoothLeAudioCodecsProviderTest
Bug: 459763325
Flag: EXEMPT TEST_ONLY
Change-Id: Iec3cee5255cf41461290e71f199052192b82ac88