Bluetooth: Add unit tests for LeAudio offload latency and ABR
This change introduces two new test cases to  to enhance coverage for
LeAudio offload IPU default implementation

The verifies the correct retrieval and parsing of, ensuring that and (including and bitmask) are accurately handled.

The focuses on ABR functionality. It confirms that values are correctly composed into when ABR is configured, and the test is conditionally executed based on the feature flag.

Bug: 442860011
Bug: 445860433
Test: atest BluetoothLeAudioCodecsProviderTest
Flag: com.android.btaudio.hal.flags.leaudio_iso_parameter_update
Change-Id: I36be6b86130983607638e3f875e0eae985737700