Make SetHciRouterBusy default-implemented virtual function
After making the SetHciRouterBusy method in TransportInterface as a pure
virtual function. This required all transport implementations to provide
an override, even if they had no specific logic to execute. To simplify
transport implementations, this patch changes SetHciRouterBusy to be a
virtual function with a default empty implementation. This makes it
optional for subclasses to override, reducing boilerplate code for
implementations that do not need to handle this notification.

Bug: 464187731
Test: Build pass
Flag: EXEMPT, no-op change.
Change-Id: I63b9f191f89971f1e973ddcbb673d68cd7a0105e