Add dataCallListUpdated to IRadioDataResponse
This change adds the `dataCallListUpdated` API to `IRadioDataResponse`.
This is done to address a race condition where the `dataCallListChanged` indication could arrive at the framework before the `setupDataCall` response, causing the framework to have an inconsistent state.

By defining this method in `IRadioDataResponse`, the HAL ensures that data call list updates are synchronized with other response callbacks, enforcing a deterministic order of events.
As per the new behavior, the system marks the data call as inactive instead of immediately removing it,waiting for a subsequent indication to finalize the removal.
Directly removing the data call from the list without changing the status to inactive would not result in data call disconnected.

The previous definition in `IRadioDataIndication` is marked as deprecated and should be replaced by the implementation in `IRadioDataResponse`.
Note that the dataCallListUpdated is still an unsolicited indication.

Test: atest CtsTelephonyTestCases
Flag: NONE AIDL changes
Bug: 438335285
Change-Id: Icbd5a97342ef9fcf9495161319690f69a2158f23