hwcryptohal_vts: Refactor HwCryptoKey VTS to use hwcrypto_thal crate
This change refactors the VtsAidlHwCryptoTests to use the
`hwcrypto_thal` crate instead of directly interacting with the
`IHwCryptoKey` service. This is part of an effort to improve the test
coverage of the `hwcrypto_thal` crate by channeling the existing VTS
tests through it.

The change currently updates only selective tests
where wrappers are already implemented in the `hwcrypto_thal` crate. The
remaining tests will be updated in subsequent changes once the
`hwcrypto_thal` crate implements all necessary wrappers.

Bug: 461380717
Test: atest VtsAidlHwCryptoTests
Change-Id: I0734e03dfd78a1591e0af60a31b0cda979820ef2