audio: Fix device proxy lifetime management for port reconfiguration
The logic of managing instances of 'BluetoothAudioPort' did not
take into account possible reconfigurations of an existing port
config. As a result, an instance of 'StreamBluetooth' could be
created without a corresponding device proxy instance.

Added a unit test replicating this scenario and fixed the code
of 'ModuleBluetoothBase' (ex-'ModuleBluetooth') to pass the test.

Bug: 463287369
Test: atest audio_module_bluetooth_tests
Test: atest VtsHalAudioCoreTargetTest
Change-Id: I7310ecd09268acb3b13f58f6d74d1697093b721a