Add BluetoothAudioSessionControl for V6 HAL
Adding another version of BluetoothAudioSessionControl.h for V6
Bluetooth Audio HAL.

Since we're adding new session types in this version, the components
using older version of Bluetooth Audio HAL will fail to build if we put
the added session types in BluetoothAudioSessionControl.h directly
because old AIDL interfaces don't have these sessions.

This is a temporary solution. We will target a refactor in the future to
fix these kinds of issues.

Bug: 424275371
Bug: 447226054
Test: m libbluetooth_audio_session_aidl
Change-Id: I4acadcacdbc9b12fed9137acd7e1c9e050084256