Introduce ITrustedHalExt
All trusted hals are expected to be extended with this interface. It is not frozen or VINTF-stable intentionally, since its purpose is to determine the default stability of the binder lib exposing the trusted HAL. This helps us try to ensure that TEEs build libbinder with the correct stability settings.

Bug: 462787437
Test: m android.hardware.security.see.ext-rust
Change-Id: I86ead36951e1539865863ebbc30686fce4de990b