Wifi: Add RangingMeasurementRole to ProximityRangingConfig
This commit introduces the RangingMeasurementRole enum to the
ProximityRangingConfig.aidl interface. This enum allows for
explicit selection of the ranging measurement role (Initiator or Responder) for a
proximity ranging session.

The new preferredRangingMeasurementRole field in the
ProximityRangingConfig parcelable provides the flexibility
to override the default role assignments, where a Seeker
defaults to Initiator and an Advertiser defaults to Responder.

Bug: 445727208
Test: atest SupplicantAidlTest
Test: atest SupplicantRttControllerAidlTest
Test: Manual - Wi-Fi RTT tests
Change-Id: I54ea871fb3cc8d015a5c902cf3f873da84ce3f1c