hwcryptohal: fixing missing placeholder_trusted_hal case
placeholder_trusted_hal did not have a definition when it was not
enabled. HwCrypto HAL expected placeholder_trusted_hal to be false
in this case insted to enable the shared secret HAL. This CL
changes the behavior to no longer need the conditional compilation.
We instead always include the ISharedSecret hal code and use an
android property to indicate if we actually need to register the
service.

Bug: 465731563
Test: atest VtsAidlSharedSecretTargetTest
Change-Id: I96869cd85489c3fa529ff39a0c41f340841772f5