Skip test for all security levels for API <= 36
Previously, the test skip was restricted to Strongbox on devices with
vendor API level 35 or lower.
The test is now also skipped for TEE KeyMint, starting from the newer
vendor API level. This is because the fix for the TEE KeyMint reference
implementation (ag/35900032) will land in the Android 17 (API level 37)
release.

Bug: 470235692
Flag: EXEMPT BUGFIX
Test: VtsAidlKeyMintTargetTest
Change-Id: If1e467fb098e9d87205aeb7233a37578c9e05af0