Provisioner should stop when the bluetooth stack has died.
A method called Stop() has been added to ChipProvisionerInterface and
its implementation to allow the HciRouter to Stop() the provisioner
if due to some reason the stack has aborted in the middle of firmware
download process. When such a 'stop' is requested, the failure to
complete firmware download now will not result in a fatal crash, rather
just an error is logged.

Bug: 465623698
Bug: 463222134
Bug: 474483347
Test: Reproduce the scenario and the bluetooth_hal should not crash
Change-Id: Ib6d6296b929efea838337d139a67ad7e3a6b3c63